agent_ai/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
langchain==0.1.0
langchain-google-genai==0.0.6
langgraph==0.0.20
pydantic==2.5.0
python-dotenv==1.0.0
pymysql==1.1.0          # Thêm MySQL driver
cryptography==41.0.7     # Cần cho pymysql
sqlalchemy==2.0.23      # ORM để làm việc với DB